+++ CHANGELOG.md
@@ -5,6 +5,9 @@
     - Support for PHP7 only.
     - Bump Symfony dependencies to `3.0`.
     - Bump PHPUnit to `6.0` and Mockery to `1.0`.
+- `Added`
+    - Added Composer script descriptions to improve Composer script documentation when runtime Composer is >= `1.6.0`. Done by [@raphaelstolt](https://github.com/raphaelstolt). See [#215](https://github.com/jonathantorres/construct/issues/215).
+    - Added Composer script descriptions for internal Composer scripts, bumps the by Construct required Composer version to `^1.6.0`. Done by [@raphaelstolt](https://github.com/raphaelstolt).
 
 #### v1.16.0 `2017-10-26`
 - `Added`